refactor!: Cleaner, more comprehensive way of handling colors
Instead of a library of constants, `Color` is now a builder class. This
gives a lot flexibility to the user: things like `Color().bold().red()`
are now possible. The downside it that it breaks the original
"user-provides-string" API. It also prevents the user from providing
custom escape sequences, by forcing him to use `Color`. This is not a
big deal, because `Color` is flexible enough now, and it also prevents
misktakes, and TextCanvas is probably not the right tool for more
complex scenarios anyway.

Also, minor refactoring of private methods: `_init_buffer()` has been
extracted out of `_clear_buffer()`. This prevents a dirty condition
(is initialized?—which was only true once during `__init__()`), and
matches the handling of the color and text buffers.