fix!: Fix `Resampling` method signatures
This is technically a breaking change, but we're fixing it quickly, so
we'll get away with a simple fix and a minor inconvenience.

Waiting for `points: &[(f64, f64)]` was a bit weird since all other
`Plot` and `Chart` methods take `x: &[f64], y: &[f64]`. This commit
changes the signatures of `downsample_mean()` and `downsample_min_max()`
accordingly.

The `points: &[(f64, f64)]` signature is still useful, and has been kept
under the `downsample_points_mean()` and `downsample_points_min_max()`
variants.

Technically, the old variants are still the "main" variants, and the new
methods are helpers to zip and unzip `x` and `y`, it's just that
API-wise it makes more sense like this.